* mouse position params are int in TUIControl, not Single.   More stable comparison in TWalkNavigator.MouseMove, and there was really no reason to keep them Single anyway (both TGLWindow and Lazarus component have them as ints, so why loose precision?)
git-svn-id: https://svn.code.sf.net/p/castle-engine/code/trunk/castle@5082 afe8dcc5-9094-451b-8773-5f495042a737

+++ source/castlegeneralmenu.pas
@@ -41,7 +41,7 @@ type
     property DrawFocused default false;
     procedure CurrentItemChanged; override;
     procedure CurrentItemSelected; override;
-    function PositionInside(const X, Y: Single): boolean; override;
+    function PositionInside(const X, Y: Integer): boolean; override;
   end;
 
 var
@@ -82,7 +82,7 @@ begin
   SoundEngine.Sound(stMenuCurrentItemSelected);
 end;
 
-function TCastleMenu.PositionInside(const X, Y: Single): boolean;
+function TCastleMenu.PositionInside(const X, Y: Integer): boolean;
 begin
   Result := true;
 end;
